Move supplementary data methods to the base QSU
