+++ app/routes/session.py
@@ -27,7 +27,7 @@ from app.helpers.template_helpers import (
 )
 from app.questionnaire import QuestionnaireSchema
 from app.questionnaire.questionnaire_schema import DEFAULT_LANGUAGE_CODE
-from app.questionnaire.questionnaire_store_updater import BaseQuestionnaireStoreUpdater
+from app.questionnaire.questionnaire_store_updater import QuestionnaireStoreUpdaterBase
 from app.questionnaire.router import Router
 from app.routes.errors import _render_error_page
 from app.services.supplementary_data import get_supplementary_data_v1
@@ -207,13 +207,10 @@ def _set_supplementary_data(
     2) Updates block and section progress to reflect any newly unlocked questions due to supplementary data list changes
     """
     router = Router(schema=schema, data_stores=questionnaire_store.data_stores)
-    base_questionnaire_store_updater = BaseQuestionnaireStoreUpdater(
+    base_questionnaire_store_updater = QuestionnaireStoreUpdaterBase(
         questionnaire_store=questionnaire_store, schema=schema, router=router
     )
-    questionnaire_store.set_supplementary_data(
-        to_set=supplementary_data,
-        base_questionnaire_store_updater=base_questionnaire_store_updater,
-    )
+    base_questionnaire_store_updater.set_supplementary_data(to_set=supplementary_data)
     base_questionnaire_store_updater.remove_dependent_blocks_and_capture_dependent_sections()
     base_questionnaire_store_updater.update_progress_for_dependent_sections()
